Improve modularity
Improved error handling when creating Memory. Now it is possible to pass
an argument (file to the text file containing the memory we want to
load) to the binary.
Improved modularity of the Chip8 by dividing execute method into helper
methods for handling values of 4 most significant bits of the op code.
Added lib.rs.