Refactor memory module with error handling
Added more robust and specific error messages.
Implemented 'new()' method  in MemeoryEntry which validates if the
address of the memory we want to write into is valid.
The 'from()' method in Memory struct now takes string slice as input,
opening file is handled by 'build()' method - this way it will be easier
to test. Removed write_byte() and write_2bytes() methods, now all
writing to memory is done via write_entry which takes MemoryEntry as
argument, this way we can be sure that the memory entry is valid, since
evalutation was already handled by MemoryEntry constructor.